pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.19.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.7",
]
